requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
diskcache==5.6.3
fastapi==0.116.2
greenlet==3.2.4
h11==0.16.0
idna==3.10
Jinja2==3.1.6
llama_cpp_python==0.3.16
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
numpy==2.3.3
pydantic==2.11.9
pydantic_core==2.33.2
PyMySQL==1.1.2
python-dotenv==1.1.1
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
